Refactor xcode connection
